chores: move util function out of plotting script
